+++ EatFitService/NutritionService/admin.py
@@ -3,7 +3,7 @@ from django.utils.functional import curry
 from django.contrib import messages
 from NutritionService.views.crowdsource_views import __create_products_from_crowdsource
 from NutritionService.models import MajorCategory, Product, MinorCategory, Allergen, NutritionFact, ErrorLog, \
-                                    CrowdsourceProduct, NotFoundLog, HealthTipp, NutrientName, \
+                                    CrowdsourceProduct, NotFoundLog, HealthTipp, NutrientName, NutriScoreFacts, \
                                     ReceiptToNutritionPartner, ReceiptToNutritionUser, Matching, DigitalReceipt, \
                                     Retailer, AdditionalImage, ImportErrorLog, MarketRegion, Ingredient
 
@@ -59,6 +59,10 @@ class NutrientInline(admin.TabularInline):
             return formset
 
 
+class NutriScoreFactsInline(admin.StackedInline):
+    model = NutriScoreFacts
+
+
 class AdditionalImageInline(admin.StackedInline):
     model = AdditionalImage
 
@@ -85,6 +89,7 @@ class ProductAdmin(admin.ModelAdmin):
     search_fields = ('product_name_de', 'gtin')
     save_as = True
     inlines = [
+        NutriScoreFactsInline,
         AdditionalImageInline,
         IngredientInline,
         AllergenInline,
@@ -146,3 +151,4 @@ admin.site.register(ReceiptToNutritionUser, ReceiptToNutritionUserAdmin)
 admin.site.register(DigitalReceipt, DigitalReceiptAdmin)
 admin.site.register(AdditionalImage)
 admin.site.register(ImportErrorLog)
+admin.site.register(NutriScoreFacts)